Fixed 64 bit issue with 'get_state' and added automatic creation and broadcast of ACCT refund transactions to the Trade package.

+++ src/Meta_Application.cpp
@@ -1531,7 +1531,7 @@ struct Meta_Application::impl : public Meta_Application_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -3142,7 +3142,8 @@ bool Meta_Application::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_Application::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Application::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -3250,7 +3251,8 @@ void Meta_Application::impl::validate( unsigned state, bool is_internal, validat
    // [<finish validate>]
 }
 
-void Meta_Application::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Application::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -4125,7 +4127,7 @@ void Meta_Application::clear( )
    p_impl->clear( );
 }
 
-void Meta_Application::validate( unsigned state, bool is_internal )
+void Meta_Application::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
