+++ src/Meta_Field.cpp
@@ -2699,7 +2699,7 @@ struct Meta_Field::impl : public Meta_Field_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -3729,7 +3729,8 @@ bool Meta_Field::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_Field::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Field::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -3899,7 +3900,8 @@ void Meta_Field::impl::validate( unsigned state, bool is_internal, validation_er
    // [<finish validate>]
 }
 
-void Meta_Field::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Field::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -5579,7 +5581,7 @@ void Meta_Field::clear( )
    p_impl->clear( );
 }
 
-void Meta_Field::validate( unsigned state, bool is_internal )
+void Meta_Field::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
