+++ src/Meta_Initial_Record_Value.cpp
@@ -414,7 +414,7 @@ struct Meta_Initial_Record_Value::impl : public Meta_Initial_Record_Value_comman
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -645,7 +645,8 @@ bool Meta_Initial_Record_Value::impl::value_will_be_provided( const string& fiel
    return false;
 }
 
-void Meta_Initial_Record_Value::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Initial_Record_Value::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -674,7 +675,8 @@ void Meta_Initial_Record_Value::impl::validate( unsigned state, bool is_internal
    // [<finish validate>]
 }
 
-void Meta_Initial_Record_Value::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Initial_Record_Value::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -997,7 +999,7 @@ void Meta_Initial_Record_Value::clear( )
    p_impl->clear( );
 }
 
-void Meta_Initial_Record_Value::validate( unsigned state, bool is_internal )
+void Meta_Initial_Record_Value::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
