+++ src/Meta_Initial_Record_Value.h
@@ -82,7 +82,7 @@ class META_INITIAL_RECORD_VALUE_DECL_SPEC Meta_Initial_Record_Value : public cla
 
    void clear( );
 
-   void validate( unsigned state, bool is_internal );
+   void validate( uint64_t state, bool is_internal );
    void validate_set_fields( std::set< std::string >& fields_set );
 
    void after_fetch( );
