+++ src/Meta_List_Type.cpp
@@ -462,7 +462,7 @@ struct Meta_List_Type::impl : public Meta_List_Type_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -744,7 +744,8 @@ bool Meta_List_Type::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_List_Type::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_List_Type::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -781,7 +782,8 @@ void Meta_List_Type::impl::validate( unsigned state, bool is_internal, validatio
    // [<finish validate>]
 }
 
-void Meta_List_Type::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_List_Type::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -1152,7 +1154,7 @@ void Meta_List_Type::clear( )
    p_impl->clear( );
 }
 
-void Meta_List_Type::validate( unsigned state, bool is_internal )
+void Meta_List_Type::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
