+++ src/Meta_Model.cpp
@@ -1230,7 +1230,7 @@ struct Meta_Model::impl : public Meta_Model_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -6933,7 +6933,8 @@ bool Meta_Model::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_Model::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Model::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -7039,7 +7040,8 @@ void Meta_Model::impl::validate( unsigned state, bool is_internal, validation_er
    // [<finish validate>]
 }
 
-void Meta_Model::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Model::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -7948,7 +7950,7 @@ void Meta_Model::clear( )
    p_impl->clear( );
 }
 
-void Meta_Model::validate( unsigned state, bool is_internal )
+void Meta_Model::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
