+++ src/Meta_Modifier.cpp
@@ -669,7 +669,7 @@ struct Meta_Modifier::impl : public Meta_Modifier_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -939,7 +939,8 @@ bool Meta_Modifier::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_Modifier::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Modifier::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -969,7 +970,8 @@ void Meta_Modifier::impl::validate( unsigned state, bool is_internal, validation
    // [<finish validate>]
 }
 
-void Meta_Modifier::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Modifier::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -1527,7 +1529,7 @@ void Meta_Modifier::clear( )
    p_impl->clear( );
 }
 
-void Meta_Modifier::validate( unsigned state, bool is_internal )
+void Meta_Modifier::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
