+++ src/Meta_Specification_Field_Action.cpp
@@ -880,7 +880,7 @@ struct Meta_Specification_Field_Action::impl : public Meta_Specification_Field_A
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -1259,7 +1259,8 @@ bool Meta_Specification_Field_Action::impl::value_will_be_provided( const string
    return false;
 }
 
-void Meta_Specification_Field_Action::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Specification_Field_Action::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -1315,7 +1316,8 @@ void Meta_Specification_Field_Action::impl::validate( unsigned state, bool is_in
    // [<finish validate>]
 }
 
-void Meta_Specification_Field_Action::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Specification_Field_Action::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -1754,7 +1756,7 @@ void Meta_Specification_Field_Action::clear( )
    p_impl->clear( );
 }
 
-void Meta_Specification_Field_Action::validate( unsigned state, bool is_internal )
+void Meta_Specification_Field_Action::validate( uint64_t state, bool is_internal )
 {
    parent_class_type::validate( state, is_internal );
    p_impl->validate( state, is_internal, &validation_errors );
