+++ src/Meta_View_Type.cpp
@@ -336,7 +336,7 @@ struct Meta_View_Type::impl : public Meta_View_Type_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -528,7 +528,8 @@ bool Meta_View_Type::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_View_Type::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_View_Type::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -565,7 +566,8 @@ void Meta_View_Type::impl::validate( unsigned state, bool is_internal, validatio
    // [<finish validate>]
 }
 
-void Meta_View_Type::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_View_Type::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -886,7 +888,7 @@ void Meta_View_Type::clear( )
    p_impl->clear( );
 }
 
-void Meta_View_Type::validate( unsigned state, bool is_internal )
+void Meta_View_Type::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
