+++ src/Meta_View_Type.h
@@ -77,7 +77,7 @@ class META_VIEW_TYPE_DECL_SPEC Meta_View_Type : public class_base
 
    void clear( );
 
-   void validate( unsigned state, bool is_internal );
+   void validate( uint64_t state, bool is_internal );
    void validate_set_fields( std::set< std::string >& fields_set );
 
    void after_fetch( );
